refactor: restructure infrastructure layer packetsx
