+++ src/main/java/br/com/matheusgusmao/incometax/auth/AuthService.java
@@ -1,14 +1,14 @@
-package br.com.matheusgusmao.incometax.authentication;
+package br.com.matheusgusmao.incometax.auth;
 
-import br.com.matheusgusmao.incometax.authentication.io.auth.AuthRequest;
-import br.com.matheusgusmao.incometax.authentication.io.auth.AuthResponse;
-import br.com.matheusgusmao.incometax.authentication.io.register.RegisterUserRequest;
-import br.com.matheusgusmao.incometax.authentication.io.register.RegisterUserResponse;
+import br.com.matheusgusmao.incometax.auth.dto.auth.AuthRequest;
+import br.com.matheusgusmao.incometax.auth.dto.auth.AuthResponse;
+import br.com.matheusgusmao.incometax.auth.dto.register.RegisterUserRequest;
+import br.com.matheusgusmao.incometax.auth.dto.register.RegisterUserResponse;
 import br.com.matheusgusmao.incometax.infra.exception.custom.EntityAlreadyExists;
 import br.com.matheusgusmao.incometax.infra.security.config.jwt.JwtService;
-import br.com.matheusgusmao.incometax.infra.security.user.model.Role;
-import br.com.matheusgusmao.incometax.infra.security.user.model.User;
-import br.com.matheusgusmao.incometax.infra.security.user.repository.UserRepository;
+import br.com.matheusgusmao.incometax.infra.security.user.Role;
+import br.com.matheusgusmao.incometax.infra.security.user.UserEntity;
+import br.com.matheusgusmao.incometax.infra.persistence.UserRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -33,7 +33,7 @@ public class AuthService {
 
         final String encodedPassword = passwordEncoder.encode(request.password());
         final UUID uuid = UUID.randomUUID();
-        final User user = User.builder().id(uuid).firstName(request.firstName()).lastName(request.lastname()).email(request.email()).password(encodedPassword).role(Role.USER).build();
+        final UserEntity user = UserEntity.builder().id(uuid).firstName(request.firstName()).lastName(request.lastname()).email(request.email()).password(encodedPassword).role(Role.USER).build();
 
         userRepository.save(user);
 
@@ -44,7 +44,7 @@ public class AuthService {
         final var authentication = new UsernamePasswordAuthenticationToken(request.username(), request.password());
         authenticationManager.authenticate(authentication);
 
-        final User user = userRepository.findByEmail(request.username()).orElseThrow(() -> new UsernameNotFoundException("Email not found: " + request.username()));
+        final UserEntity user = userRepository.findByEmail(request.username()).orElseThrow(() -> new UsernameNotFoundException("Email not found: " + request.username()));
         final String token = jwtService.generateToken(user);
 
         return new AuthResponse(token);
