+++ src/main/java/br/com/matheusgusmao/incometax/infra/persistence/UserRepository.java
@@ -1,0 +1,11 @@
+package br.com.matheusgusmao.incometax.infra.persistence;
+
+import br.com.matheusgusmao.incometax.infra.security.user.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface UserRepository extends JpaRepository<UserEntity, UUID> {
+    Optional<UserEntity> findByEmail(String email);
+}
