+++ src/main/java/br/com/matheusgusmao/incometax/infra/security/user/Role.java
@@ -1,0 +1,6 @@
+package br.com.matheusgusmao.incometax.infra.security.user;
+
+public enum Role {
+    USER,
+    ADMIN
+}
