+++ src/main/java/br/com/matheusgusmao/incometax/infra/security/user/UserEntity.java
@@ -1,4 +1,4 @@
-package br.com.matheusgusmao.incometax.infra.security.user.model;
+package br.com.matheusgusmao.incometax.infra.security.user;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -10,18 +10,19 @@ import org.springframework.security.core.userdetails.UserDetails;
 import java.sql.Types;
 import java.util.Collection;
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
-@Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
 @Table(name = "app_user")
-public class User implements UserDetails {
+public class UserEntity implements UserDetails {
     @Id
     @JdbcTypeCode(Types.VARCHAR)
     @NonNull @Column(nullable = false, unique = true)
+    @Getter
     private UUID id;
     @NonNull @Column(nullable = false)
     private String firstName;
@@ -37,6 +38,18 @@ public class User implements UserDetails {
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
         return List.of(new SimpleGrantedAuthority(role.name()));
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (o == null || getClass() != o.getClass()) return false;
+        UserEntity user = (UserEntity) o;
+        return Objects.equals(getId(), user.getId());
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(getId());
     }
 
     @Override
