--- src/main/java/br/com/matheusgusmao/incometax/infra/security/user/model/Role.java
@@ -1,6 +1,0 @@
-package br.com.matheusgusmao.incometax.infra.security.user.model;
-
-public enum Role {
-    USER,
-    ADMIN
-}
