--- src/main/java/br/com/matheusgusmao/incometax/infra/security/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package br.com.matheusgusmao.incometax.infra.security.user.repository;
-
-import br.com.matheusgusmao.incometax.infra.security.user.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface UserRepository extends JpaRepository<User, UUID> {
-    Optional<User> findByEmail(String email);
-}
